HPS-27 refactor: move user structures to separate user folders
