+++ src/main/java/com/example/colaba/dto/user/UpdateUserRequest.java
@@ -1,4 +1,4 @@
-package com.example.colaba.dto;
+package com.example.colaba.dto.user;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Size;
