+++ src/main/java/com/example/colaba/dto/user/UserScrollResponse.java
@@ -1,4 +1,4 @@
-package com.example.colaba.dto;
+package com.example.colaba.dto.user;
 
 import lombok.Getter;
 import lombok.Setter;
